requirements.txt
fastapi
uvicorn
boto3
pymysql
python-dotenv